Split cwrapper into 2 parts to avoid macro for dependencies
